# Improved ParkTakeStrategy