It includes 2 improvements for park take strategies:
- single-consumer: uses field updater to avoid AtomicReference instance
- multi-consumer: uses intrinsic locks to avoid garbage on contention
and avoided locking while checking waiting threads