@franz1981 2019-01-13
Using my tiny, but quiet, box I have gotten these numbers with `McParkTakeStrategy`:
```
MASTER (with ReentrantLock):
Benchmark                 (burstSize)  (consumerCount)  (qCapacity)  (qType)  (warmup)  Mode  Cnt      Score      Error  Units
QueueBurstCost.burstCost            1                1       132000      608      true  avgt   20    196.313 ±    4.873  ns/op
QueueBurstCost.burstCost          100                1       132000      608      true  avgt   20  10302.968 ± 1110.205  ns/op


MASTER with intrinsic lock:
Benchmark                 (burstSize)  (consumerCount)  (qCapacity)  (qType)  (warmup)  Mode  Cnt     Score     Error  Units
QueueBurstCost.burstCost            1                1       132000      608      true  avgt   20   196.327 ±   6.058  ns/op
QueueBurstCost.burstCost          100                1       132000      608      true  avgt   20  9722.554 ± 148.198  ns/op


PR with intrinsic lock:

Benchmark                 (burstSize)  (consumerCount)  (qCapacity)  (qType)  (warmup)  Mode  Cnt     Score     Error  Units
QueueBurstCost.burstCost            1                1       132000      608      true  avgt   20   195.678 ±   6.314  ns/op
QueueBurstCost.burstCost          100                1       132000      608      true  avgt   20  5574.261 ± 146.196  ns/op

PR with ReentrantLock:

Benchmark                 (burstSize)  (consumerCount)  (qCapacity)  (qType)  (warmup)  Mode  Cnt     Score     Error  Units
QueueBurstCost.burstCost            1                1       132000      608      true  avgt   20   192.659 ±   6.307  ns/op
QueueBurstCost.burstCost          100                1       132000      608      true  avgt   20  5519.689 ± 142.494  ns/op
```
it seems that on the single-producer single-consumer usage of this strategy the improvement is dramatic: 
```
10302.968 ± 1110.205  ns/op VS 5574.261 ± 146.196  ns/op
```

The weird thing (need to check the ASM, probably) is that the original `ScParkTakeStrategy`is performing worse then `McParkTakeStrategy` for `burstSize=100`:
```
Benchmark                 (burstSize)  (consumerCount)  (qCapacity)  (qType)  (warmup)  Mode  Cnt      Score      Error  Units
QueueBurstCost.burstCost            1                1       132000      608      true  avgt  100    197.587 ±    3.148  ns/op
QueueBurstCost.burstCost          100                1       132000      608      true  avgt  100   7724.747 ±  223.752  ns/op
```
FYI the new born Mpsc blocking array q:
```
Benchmark                 (burstSize)  (consumerCount)  (qCapacity)  (qType)  (warmup)  Mode  Cnt      Score      Error  Units
QueueBurstCost.burstCost            1                1       132000       72      true  avgt  100    197.071 ±    3.232  ns/op
QueueBurstCost.burstCost          100                1       132000       72      true  avgt  100   5637.344 ±  259.194  ns/op
```
I was expecting an even better behaviour TBH, but probably I need to play with the producers number/consumers number/DELAY_CONSUMER/DELAY_PRODUCER.